pyproject.toml
[tool.poetry]
name = "financial-analyzing-planning"
version = "0.1.0"
description = "Creating Multi Agent Financial Planner"
authors = ["Z"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
crewai = "^0.11.0"
duckduckgo-search = "^4.4.2"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.3"

[tool.pyright]
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
